Separated unsafe code for App Engine support.
